requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
et-xmlfile==1.1.0
Flask==2.2.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.1
openai==0.22.1
openpyxl==3.0.10
pandas==1.4.3
pandas-stubs==1.4.3.220801
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
six==1.16.0
tqdm==4.64.0
twilio==7.12.0
types-pytz==2022.1.2
urllib3==1.26.11
Werkzeug==2.2.1
zipp==3.8.1
